refactor(trackers): group Gazelle trackers
