+++ src/trackersetup.py
@@ -1417,17 +1417,17 @@ tracker_class_map: Any = LazyTrackerDict(
     {
         "1PTBA": ("src.trackers.NEXUSPHP.oneptba", "OnePTBA"),
         "AITHER": ("src.trackers.UNIT3D.aither", "Aither"),
-        "ALPHARATIO": ("src.trackers.alpharatio", "AlphaRatio"),
+        "ALPHARATIO": ("src.trackers.GAZELLE.alpharatio", "AlphaRatio"),
         "AMIGOSSHARE": ("src.trackers.amigosshare", "AmigosShare"),
-        "ANTHELION": ("src.trackers.anthelion", "Anthelion"),
+        "ANTHELION": ("src.trackers.GAZELLE.anthelion", "Anthelion"),
         "ASIANCINEMA": ("src.trackers.UNIT3D.asiancinema", "AsianCinema"),
         "AVISTAZ": ("src.trackers.AVISTAZ.avistaz", "AvistaZ"),
         "BEYONDHD": ("src.trackers.beyondhd", "BEYONDHD"),
         "BITHDTV": ("src.trackers.bithdtv", "BitHDTV"),
         "BITPORN": ("src.trackers.UNIT3D.bitporn", "BitPorn"),
-        "BJSHARE": ("src.trackers.bjshare", "BJShare"),
+        "BJSHARE": ("src.trackers.GAZELLE.bjshare", "BJShare"),
         "BLUTOPIA": ("src.trackers.UNIT3D.blutopia", "Blutopia"),
-        "BRASILTRACKER": ("src.trackers.brasiltracker", "BrasilTracker"),
+        "BRASILTRACKER": ("src.trackers.GAZELLE.brasiltracker", "BrasilTracker"),
         "BROADCASTHENET": ("src.trackers.broadcasthenet", "BroadcasTheNet"),
         "CAPYBARABR": ("src.trackers.UNIT3D.capybarabr", "CapybaraBR"),
         "CATHODERAYTUBE": ("src.trackers.cathoderaytube", "CathodeRayTube"),
@@ -1443,7 +1443,7 @@ tracker_class_map: Any = LazyTrackerDict(
         "FILELIST": ("src.trackers.filelist", "FileList"),
         "FLOOD": ("src.trackers.flood", "Flood"),
         "FUNFILE": ("src.trackers.funfile", "FunFile"),
-        "GREATPOSTERWALL": ("src.trackers.greatposterwall", "GreatPosterWall"),
+        "GREATPOSTERWALL": ("src.trackers.GAZELLE.greatposterwall", "GreatPosterWall"),
         "HAWKEUNO": ("src.trackers.UNIT3D.hawkeuno", "HawkeUno"),
         "HDBITS": ("src.trackers.hdbits", "HDBits"),
         "HDSPACE": ("src.trackers.hdspace", "HDSpace"),
@@ -1464,13 +1464,13 @@ tracker_class_map: Any = LazyTrackerDict(
         "MAKINGOFF": ("src.trackers.makingoff", "MakingOff"),
         "MIDNIGHTSCENE": ("src.trackers.UNIT3D.midnightscene", "MidnightScene"),
         "MTEAM": ("src.trackers.mteam", "MTeam"),
-        "NEBULANCE": ("src.trackers.nebulance", "Nebulance"),
+        "NEBULANCE": ("src.trackers.GAZELLE.nebulance", "Nebulance"),
         "NORDICQUALITY": ("src.trackers.UNIT3D.nordicquality", "NordicQuality"),
         "NZBGEEK": ("src.trackers.USENET.nzbgeek", "NZBGeek"),
         "OLDTOONSWORLD": ("src.trackers.UNIT3D.oldtoonsworld", "OldToonsWorld"),
         "ONLYENCODES": ("src.trackers.UNIT3D.onlyencodes", "OnlyEncodes"),
-        "ORPHEUS": ("src.trackers.orpheus", "Orpheus"),
-        "PASSTHEPOPCORN": ("src.trackers.passthepopcorn", "PassThePopcorn"),
+        "ORPHEUS": ("src.trackers.GAZELLE.orpheus", "Orpheus"),
+        "PASSTHEPOPCORN": ("src.trackers.GAZELLE.passthepopcorn", "PassThePopcorn"),
         "PEERGARDEN": ("src.trackers.UNIT3D.peergarden", "PeerGarden"),
         "POLISHTORRENT": ("src.trackers.UNIT3D.polishtorrent", "PolishTorrent"),
         "PORTUGAS": ("src.trackers.UNIT3D.portugas", "Portugas"),
