+++ tests/test_greatposterwall_image_host.py
@@ -5,7 +5,7 @@
 import asyncio
 
 from src.meta import Meta
-from src.trackers.greatposterwall import GreatPosterWall
+from src.trackers.GAZELLE.greatposterwall import GreatPosterWall
 
 
 class _Response:
@@ -36,7 +36,7 @@ class _Client:
 
 
 def test_greatposterwall_rehosts_only_unapproved_urls(monkeypatch):
-    monkeypatch.setattr("src.trackers.greatposterwall.httpx.AsyncClient", _Client)
+    monkeypatch.setattr("src.trackers.GAZELLE.greatposterwall.httpx.AsyncClient", _Client)
     tracker = GreatPosterWall({"DEFAULT": {"tmdb_api": "test"}, "TRACKERS": {"GREATPOSTERWALL": {"api_key": "test-key"}}})
     meta = Meta(
         image_list=[
